% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_fish.R
\name{compute_fish}
\alias{compute_fish}
\title{Fish Diversity}
\usage{
compute_fish(fish_vector)
}
\arguments{
\item{species}{vector of species}
}
\value{
list with the following items
\describe{
\item {Name of the most common species}
\item {Name of the rarest species}
\item {Total number of species}
}
}
\description{
This function takes a vector of fish names and returns the most common species, the rarest species and the total number of species
}
