% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_df.R
\docType{data}
\name{fish_df}
\alias{fish_df}
\title{Fish Data}
\format{
A data frame with 5 rows and 2 variables:
\itemize{
\item name of 5 different types of fish
\item number of different fish types
}
}
\usage{
fish_df
}
\description{
Data on fish in a hypothetical body of water
}
\keyword{datasets}
