% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_df.R
\docType{data}
\name{inputs_df}
\alias{inputs_df}
\title{Sun and Water for Plants}
\format{
A data frame with 52 rows and 2 variables:
\itemize{
\item sun (hours per week)
\item water (waterings per week)
}
}
\usage{
inputs_df
}
\description{
Data on sun and water input values for plant growth
}
\keyword{datasets}
