% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_df.R
\docType{data}
\name{plant_df}
\alias{plant_df}
\title{Environmental Conditions for Plant Growth}
\format{
A data frame with 52 rows and 2 variables:
\itemize{
\item sun (hours per week)
\item water (watering events per week)
}
}
\usage{
plant_df
}
\description{
Dataframe of sun and water values for input into the compute_plant_growth function. Values in the 'sun' column represent amount of sunlight given to a plant per week. Values in the 'water' column represent number of plant-watering events per week.
}
\keyword{datasets}
