% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_plant_growth.R
\name{compute_plant_growth}
\alias{compute_plant_growth}
\title{Growth of Plant over time}
\usage{
compute_plant_growth(conditions, pot = 1)
}
\arguments{
\item{conditions}{dataframe containing 'sun' column with number of sun hours per week (n) and 'water' column containing number of watering events per week (n)}

\item{pot}{size of pot (diameter in feet) the plant is growing in}
}
\value{
list with the following items
\itemize{
\item {height of plant (mm)}
\item {number of leaves (n)}
}
}
\description{
Compute the growth of a plant over time based on environmental conditions.
}
\author{
Kristin Art, Sage Davis, Daija Odom
}
