% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_df.R
\docType{data}
\name{fish_df}
\alias{fish_df}
\title{Fish Catch Data}
\format{
A vector with 20 rows and 1 variable:
\itemize{
\item fish species caught (cod, salmon, tuna, steelhead, or shark)
}
}
\usage{
fish_df
}
\description{
Vector of example data of one fish catch with multiple species for input into the compute_fish function. Each row represents one individual fish that was included in the catch. Each cell includes a character of the caught individuals' species name.
}
\keyword{datasets}
